cassandra-rest-api/requirements.txt
numpy==1.19.5
pandas==1.1.5
requests==2.26.0
urllib3==1.26.6
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy<1.4
cassandra-driver==3.25.0
yfinance==0.1.63
matplotlib==3.3.4
plotly==5.1.0
scikit-learn==0.24.2
tensorflow==2.5.0